requirements.txt
pygame==2.0.0.dev12
PyGLM
PyOpenGL
esper